+++ app.js
@@ -1,83 +1,64 @@
-var app = angular.module('flapperNews', ['ui.router']);
-app.factory('posts', [function(){
-  var o = {
-    posts: []
-  };
-  return o;
-}]);
+var express = require('express');
+var path = require('path');
+var favicon = require('serve-favicon');
+var logger = require('morgan');
+var cookieParser = require('cookie-parser');
+var bodyParser = require('body-parser');
 
-app.controller('MainCtrl', [
-    '$scope',
-    'posts',
-    function($scope, posts){
-        $scope.posts = posts.posts;
-        
-        //add a new post (link is optional)
-        $scope.addPost = function(){
-          //prevent user from entering blank title
-          if(!$scope.title || $scope.title === "") {
-              return;
-          }
-          
-          $scope.posts.push({title: $scope.title, link: $scope.link, upvotes: 0,
-            comments: [
-              {author: 'Joe', body: 'Cool post!', upvotes: 0},
-              {author: 'Bob', body: 'Great idea but everything is wrong!', upvotes: 0}
-            ]
-          });  
-          $scope.title = "";
-          $scope.link = "";
-        };
-        
-        //upvote a post
-        $scope.incrementUpvotes = function(post){
-          post.upvotes += 1;  
-        };
-    }
-]);
+var routes = require('./routes/index');
+var users = require('./routes/users');
 
-//will control posts and comments
-app.controller('PostsCtrl', [
-'$scope',
-'$stateParams',
-'posts',
-function($scope, $stateParams, posts){ //has access to posts factory
-    $scope.posts = posts.posts[$stateParams.id];
-    
-    //add a comment to a post
-    $scope.addComment = function(){
-      if($scope.body === '') {
-        return;
-      }
-      $scope.post.comments.push({
-        body: $scope.body,
-        author: 'user',
-        upvotes: 0
-      });
-      $scope.body = '';
-    };
-    
-}]);
+//Connect to MongoDB Mongoose instance
+var mongoose = require("mongoose");
+mongoose.connect("mongodb://kshen3778:qazwsx1703@" + process.env.IP + ":" + process.env.PORT + "/news");
 
-app.config([
-'$stateProvider',
-'$urlRouterProvider',
-function($stateProvider, $urlRouterProvider){
-    
-    //home state
-    $stateProvider.state('home', {
-      url: '/home',
-      templateUrl: '/home.html',
-      controller: 'MainCtrl'
+var app = express();
+
+// view engine setup
+app.set('views', path.join(__dirname, 'views'));
+app.set('view engine', 'ejs');
+
+// uncomment after placing your favicon in /public
+//app.use(favicon(path.join(__dirname, 'public', 'favicon.ico')));
+app.use(logger('dev'));
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: false }));
+app.use(cookieParser());
+app.use(express.static(path.join(__dirname, 'public')));
+
+app.use('/', routes);
+app.use('/users', users);
+
+// catch 404 and forward to error handler
+app.use(function(req, res, next) {
+  var err = new Error('Not Found');
+  err.status = 404;
+  next(err);
+});
+
+// error handlers
+
+// development error handler
+// will print stacktrace
+if (app.get('env') === 'development') {
+  app.use(function(err, req, res, next) {
+    res.status(err.status || 500);
+    res.render('error', {
+      message: err.message,
+      error: err
     });
-    
-    //posts and comments
-    $stateProvider.state('posts', {
-      url: '/posts/{id}',
-      templateUrl: '/posts.html',
-      controller: 'PostsCtrl'
-    });
-    
-    $urlRouterProvider.otherwise('home');
-    
-}]);
+  });
+}
+
+// production error handler
+// no stacktraces leaked to user
+app.use(function(err, req, res, next) {
+  res.status(err.status || 500);
+  res.render('error', {
+    message: err.message,
+    error: {}
+  });
+});
+
+
+module.exports = app;
