+++ npm-debug.log
@@ -1,0 +1,28 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/home/ubuntu/.nvm/v0.10.31/bin/node',
+1 verbose cli   '/home/ubuntu/.nvm/v0.10.31/bin/npm',
+1 verbose cli   'start' ]
+2 info using npm@1.4.23
+3 info using node@v0.10.31
+4 verbose run-script [ 'prestart', 'start', 'poststart' ]
+5 info prestart workspace@0.0.0
+6 info start workspace@0.0.0
+7 verbose unsafe-perm in lifecycle true
+8 info workspace@0.0.0 Failed to exec start script
+9 error workspace@0.0.0 start: `node ./bin/www`
+9 error Exit status 8
+10 error Failed at the workspace@0.0.0 start script.
+10 error This is most likely a problem with the workspace package,
+10 error not with npm itself.
+10 error Tell the author that this fails on your system:
+10 error     node ./bin/www
+10 error You can get their info via:
+10 error     npm owner ls workspace
+10 error There is likely additional logging output above.
+11 error System Linux 3.14.13-c9
+12 error command "/home/ubuntu/.nvm/v0.10.31/bin/node" "/home/ubuntu/.nvm/v0.10.31/bin/npm" "start"
+13 error cwd /home/ubuntu/workspace
+14 error node -v v0.10.31
+15 error npm -v 1.4.23
+16 error code ELIFECYCLE
+17 verbose exit [ 1, true ]
