+++ views/index.ejs
@@ -1,0 +1,87 @@
+<html>
+    <head>
+        <title>Flapper News</title>
+        <link href="https://maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css" rel="stylesheet">
+        <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.3.10/angular.min.js"></script>
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/angular-ui-router/0.2.10/angular-ui-router.js"></script>
+        <script src="/javascripts/angularApp.js"></script>
+        <style> .glyphicon-thumbs-up {cursor:pointer} </style>
+    </head>
+    
+    <body ng-app="flapperNews" >
+        <div class="row">
+            <div class="col-md-6 col-md-offset-3">
+                <ui-view></ui-view>
+            </div>
+        </div>
+        
+         <!-- Inline home.html template -->
+        <script type="text/ng-template" id="/home.html">
+            <div class="page-header">
+                <h1>Flapper News</h1>
+            </div>
+            
+            <div ng-repeat="post in posts | orderBy: '-upvotes'">
+                <span class="glyphicon glyphicon-thumbs-up" ng-click="incrementUpvotes(post)"></span>
+                {{post.upvotes}}
+                <span style="font-size:20px; margin-left:10px">
+                    <a ng-show="post.link" href="{{post.link}}">
+                        {{post.title}}
+                    </a>
+                    <span ng-hide="post.link">
+                        {{post.title}}
+                    </span>
+                    <span>
+                        <a href="#/posts/{{$index}}">Comments</a>
+                    </span>
+                </span>
+            </div>
+                    
+            <form ng-submit="addPost()" style="margin-top:30px;">
+                <h3>Add a new post</h3>
+                        
+                <div class="form-group">
+                    <input type="text" class="form-control" placeholder="Title" ng-model="title"></input>
+                </div>
+                <div class="form-group">
+                    <input type="text" class="form-control" placeholder="Link" ng-model="link"></input>
+                </div>
+                <button type="submit" class="btn btn-primary">Post</button>
+            </form>
+            
+        </script>
+        
+        <script type="text/ng-template" id="/posts.html">
+            <div class="page-header">
+                <h3>
+                    <a ng-show="post.link" href="{{post.link}}">
+                        {{post.title}}
+                    </a>
+                    <span ng-hide="post.link">
+                        {{post.title}}
+                    </span>
+                </h3>
+            </div>
+            
+            <div ng-repeat="comment in post.comments | orderBy: '-upVotes">
+                 <span class="glyphicon glyphicon-thumbs-up" ng-click="incrementUpvotes(post)"></span>
+                {{comment.upvotes}} - by {{comment.author}}
+                <span style="font-size:20px; margin-left:10px">
+                    {{comment.body}}
+                </span>
+            </div>
+            
+            <form ng-submit="addComment" style="margin-top: 30px;">
+                <h3>Add a new comment</h3>
+                
+                <div class="form-group">
+                    <input type="text" class="form-control" placeholder="Comment" ng-model="body"></input>
+                    
+                </div>
+                <button type="submit" class="btn btn-primary">Post</button>
+            </form>
+        </script>
+        
+        
+    </body>
+</html>
